sqlbase: move index encoding functions to their own file
Release note: None
